+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.2.22 - Unreleased
 
+### Added
+
+* `path_sanatize` to `kwcoco.ChannelSpec` to make a path-safe suffix for a filename
+* Add experimental `_dataset_id` function to `kwcoco.CocoDataset` for unique
+  but human interpretable dataset identifiers.
+
 
 ## Version 0.2.21 - Released 2022-02-15